requirements.txt
urllib3==1.24.1
django-allauth==0.32.0
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-extensions==1.7.9
django-filter==1.0.4
django-filters==0.2.1
django-rest-auth==0.9.1
django-rest-framework==0.1.0
djangorestframework==3.7.7